refactor(distri-core): extract UniversalAgentTool into dedicated module
Pure module extraction, no logic change. Prepares for Phase 4b unification
of CallMode (fork/offload/transfer) and deletion of AgentTool and
TransferToAgentTool.

- New file: server/distri-core/src/tools/universal_agent.rs holding
  UniversalAgentTool + CallAgentInput + is_agent_accessible +
  resolve_code_agent + ALWAYS_AVAILABLE_BUILTINS.
- spawn_background_execution bumped to pub(crate) (needed by the new
  module in Phase 4b).
- Test imports updated from tools::builtin:: to tools::universal_agent::.
