+++ server/distri-core/src/tools/mod.rs
@@ -31,10 +31,12 @@ pub mod send_message;
 pub mod simulator;
 pub mod skill_script;
 pub mod tool_search;
+pub mod universal_agent;
 pub use builtin::{
     get_builtin_tools, AgentTool, ConsoleLogTool, DistriExecuteCodeTool, FinalTool,
-    TransferToAgentTool, UniversalAgentTool,
+    TransferToAgentTool,
 };
+pub use universal_agent::UniversalAgentTool;
 pub use inject_env::InjectConnectionEnvTool;
 pub use send_message::SendMessageTool;
 pub use tool_search::ToolSearchTool;
